Pacemaker: set force-stop to false
There is no reason to not try a normal stop before forcing one. It seems
better to try to stop cleanly all guests that support it.

Pacemaker documentation states: "The default
behavior is to resort to a forceful shutdown only after a graceful
shutdown attempt has failed. You should only set this to true if
your virtual domain (or your virtualization backend) does not support
graceful shutdown."
https://github.com/ClusterLabs/resource-agents/blob/8483b12455eb2aef272c13421eb2f8a7fce8d246/heartbeat/VirtualDomain#L116

The stop timeout is now set to 15s, so that the forceful shutdown occurs
quite rapidly if the clean shutdown does not work.

Signed-off-by: insatomcat <florent.carli@rte-france.com>