stop command: add force option
The force option was not implemented yet for stopping a VM.
It can be called on the command line with -f or --force.
If vm_manager is used in non-cluster mode, the command "stop --force"
will call libvirt function destroy() instead of shutdown().
The option remains unimplemented for cluster mode : it will still call
Pacemaker.stop()

Change-Id: I68ee1f9961970e601ad36ea560cd7b43f2270cb1
Signed-off-by: Thibault Sourdin <thibault.sourdin@savoirfairelinux.com>