add extra disk support to vm_manager
Extra disks are empty disks that can be created and attached to a VM.

The extra disks are created with the --extra-disks option of the create
command. The option can be used multiple times to create multiple extra
disks. This option takes the size of the disk in MB or with a suffix (K,
M, G, T). The clone command also supports the --clear_extra_data option
to clear the content of the extra disks when cloning a VM.

The extra disks are created in the same RBD group as the system disk and
are also snapshotted when the VM is snapshotted.

Signed-off-by: Mathieu Dupré <mathieu.dupre@savoirfairelinux.com>