rbd_manager: improve import disk speed
This patch adds the -W option to the qemu-img convert command. This
option enables the out of order writes, which can improve the import
speed of the disk up to 2.3x. The -W option can only be used with RAW
disk images as the destination format (which is the case here).

Signed-off-by: Mathieu Dupré <mathieu.dupre@savoirfairelinux.com>