vm_manager_cluster: fix live_migration option
The live_migration option is set to the value of the command line
argument enable_live_migration. This argument's default value is False,
thus live_migration will always be in the vm_options dictionary.

Fix condition to add the "_live_migration" metadata which should check
if vm_options["live_migration"] is True rather than checking if it
exists.

Signed-off-by: Elinor Montmasson <elinor.montmasson@savoirfairelinux.com>