vm_manager: fix error handling when cloning xml configuraton
If the "_base_xml" metadata from the src disk doesn't exist, an
exception is raised. If it exists, a value is returned and the element
"base_xml" will be in vm_options. In any case, the following if block
can never be entered.

Remove this unused if block and if the metadata doesn't exist, log the
error and raise again the KeyError.

Signed-off-by: Elinor Montmasson <elinor.montmasson@savoirfairelinux.com>