pacemaker: add support for custom Pacemaker configuration
Add the ability to specify custom Pacemaker "params", "meta" and
"utilization" configuration for a VM.

There are some cases where "crm_config_cmd" is not enough to configure
a VM in Pacemaker. "crm_config_cmd" will call crm command to add custom
configuration after the VM is added to the cluster. This can be an
issue when we need to configure utilization.

Add "clear" options to remove custom configuration from the VM during
the clone operation.

Add new arguments to the command line interface:

- For create and clone commands:
    - --pacemaker-meta: Set a number of key-value pacemaker "meta"
    - --pacemaker-params: Set a number of key-value pacemaker "params"
    - --pacemaker-utilization: Set a number of key-value pacemaker
        "utilization"
- For clone command:
    - --clear-pacemaker-meta: Do not keep custom pacemaker meta
    - --clear-pacemaker-params: Do not keep custom pacemaker params
    - --clear-pacemaker-utilization: Do not keep custom pacemaker
        utilization

If during the clone operation, the custom configuration is specified
and the cloned VM has already custom configuration, the two
configurations will be merged. If two keys are the same, the new value
will overwrite the old one.

Signed-off-by: Mathieu Dupré <mathieu.dupre@savoirfairelinux.com>