pacemaker: add support for custom Pacemaker configuration
Add the ability to specify custom Pacemaker "params", "meta" and
"utilization" configuration for a VM.

There are some cases where "crm_config_cmd" is not enough to configure
a VM in Pacemaker. "crm_config_cmd" will call crm command to add custom
configuration after the VM is added to the cluster. This can be an
issue we need to configure utilization.

Signed-off-by: Mathieu Dupré <mathieu.dupre@savoirfairelinux.com>