vm_manager_cmd: add console command
The console command allows opening a console on a VM. It is available
in both cluster and standalone mode on the command line only.

It uses the virsh console command to open a console on the VM. On
cluster mode, it first finds the hypervisor where the VM is running
using the Pacemaker helper.

To be used on cluster mode it required to have an SSH connection to the
hypervisor where the VM is running. The user can be specified with the
--ssh-user option.

Signed-off-by: Mathieu Dupré <mathieu.dupre@savoirfairelinux.com>